separate bundle tests
